cleanup examples

+++ examples/example-f64/src/main.rs
@@ -6,5 +6,5 @@ fn main() {
     println!("Hello, world!");
     let mut dsp = Volume::new();
     dsp.init(44_100);
-    dsp.set_param(ParamIndex(0), 10.0 as f64);
+    dsp.set_param(ParamIndex(0), 10.0_f64);
 }
